Simplify Now Playing wiring
