+++ app/components/player/nowplaying/nowplaying_other.go
@@ -1,21 +1,11 @@
 //go:build !darwin
 
+// Stubs for non-Darwin builds; every exported function is a no-op.
 package nowplaying
 
-// Configure is a no-op on non-Darwin platforms.
 func Configure(info Info, handlers Handlers) {}
-
-// SetTextMetadata is a no-op on non-Darwin platforms.
-func SetTextMetadata(info Info) {}
-
-// SetState is a no-op on non-Darwin platforms.
-func SetState(state State) {}
-
-// SetPosition is a no-op on non-Darwin platforms.
-func SetPosition(positionUs int64) {}
-
-// SetArtwork is a no-op on non-Darwin platforms.
-func SetArtwork(data []byte) {}
-
-// Clear is a no-op on non-Darwin platforms.
-func Clear() {}
+func SetTextMetadata(info Info)              {}
+func SetState(state State)                   {}
+func SetPosition(positionUs int64)           {}
+func SetArtwork(data []byte)                 {}
+func Clear()                                 {}
