+++ app/components/player/player_nowplaying_other.go
@@ -1,0 +1,13 @@
+//go:build !darwin
+
+package player
+
+import (
+	"context"
+
+	"github.com/0skillallluck/scanline/app/sources"
+)
+
+// fetchAndPushArtwork is a no-op on non-Darwin: nowplaying.SetTextMetadata /
+// SetArtwork are stubs there, so we'd just be issuing wasted HTTP fetches.
+func fetchAndPushArtwork(ctx context.Context, src sources.Source, params PlayerParams) {}
